loosen requirement to manifolds.
