+++ src/functions/manifold_functions.jl
@@ -35,38 +35,21 @@ function mid_point!(M::Sphere, y, p, q, x)
     return y
 end
 
-@doc raw"""
-    reflect(M, f, x)
-
-reflect the point `x` from the manifold `M` at the point `f(x)` of the
-function $f\colon \mathcal M → \mathcal M$, i.e.,
-
-````math
-    \operatorname{refl}_f(x) = \operatorname{refl}_{f(x)}(x),
-````
-see also [`reflect`](@ref reflect(M::Manifold, p, x))`(M,p,x)`.
-"""
-
-reflect(M::Manifold, pr::Function, x) = reflect(M, pr(x), x)
-
-@doc raw"""
-    reflect(M, p, x)
-
-reflect the point `x` from the manifold `M` at point `x`, i.e.
-
-````math
-    \operatorname{refl}_p(x) = \exp_p(-\log_p x).
-````
-where exp and log denote the exponential and logarithmic map on `M`.
-"""
-reflect(M::Manifold, p, x) = exp(M, p, -log(M, p, x))
-
-@doc raw"""
-    sym_rem(x,[T=π])
-
-Compute symmetric remainder of `x` with respect to the interall 2*`T`, i.e.
-`(x+T)%2T`, where the default for `T` is $π$
-"""
-function sym_rem(x::N, T=π) where {N<:Number}
-    return (x ≈ T ? convert(N, -T) : rem(x, convert(N, 2 * T), RoundNearest))
+function prox_TV2(::Euclidean, λ, pointTuple::Tuple{T,T,T}, p::Int=1) where {T}
+    w = @SVector [1.0, -2.0, 1.0]
+    x = SVector(pointTuple)
+    if p == 1 # Example 3.2 in Bergmann, Laus, Steidl, Weinmann, 2014.
+        m = min.(Ref(λ), abs.(x .* w) / (dot(w, w)))
+        s = sign.(sum(x .* w))
+        return x .- m .* s .* w
+    elseif p == 2 # Theorem 3.6 ibd.
+        t = λ * sum(x .* w) / (1 + λ * dot(w, w))
+        return x .- t .* w
+    else
+        throw(
+            ErrorException(
+                "Proximal Map of TV2(Euclidean,λ,pT,p) not implemented for p=$(p) (requires p=1 or 2)",
+            ),
+        )
+    end
 end
